+++ src/Rule/AliasRule.php
@@ -65,7 +65,7 @@ class AliasRule extends AbstractRule
     /**
      * @inheritdoc
      */
-    public function cleanConstraint(mixed $value): mixed
+    protected function cleanConstraint(mixed $value): mixed
     {
         $value = parent::cleanConstraint($value);