+++ src/Rule/DateRule.php
@@ -95,7 +95,7 @@ class DateRule extends AbstractRule
     /**
      * @inheritdoc
      */
-    public function cleanConstraint(mixed $value): mixed
+    protected function cleanConstraint(mixed $value): mixed
     {
         $isDateTime = false;