+++ src/Rule/DateTimeRule.php
@@ -96,7 +96,7 @@ class DateTimeRule extends AbstractRule
     /**
      * @inheritdoc
      */
-    public function cleanConstraint(mixed $value): mixed
+    protected function cleanConstraint(mixed $value): mixed
     {
         $isDateTime = false;