+++ src/Rule/RequiredRule.php
@@ -1,6 +1,7 @@
 <?php
 namespace Pyncer\Validation\Rule;
 
+use Closure;
 use Pyncer\Exception\InvalidArgumentException;
 use Pyncer\Validation\Rule\RuleInterface;
 use Stringable;
@@ -55,7 +56,7 @@ class RequiredRule implements RuleInterface
             $value = trim($value);
         }
 
-        if (in_array($value, $this->empty, true)) {
+        if ($this->isEmptyValue($value)) {
             return false;
         }
 
@@ -92,5 +93,20 @@ class RequiredRule implements RuleInterface
     public function getError(): ?string
     {
         return 'required';
+    }
+
+    protected function isEmptyValue(mixed $value): bool
+    {
+        foreach ($this->empty as $emptyValue) {
+            if ($emptyValue instanceof Closure) {
+                if (call_user_func($emptyValue, $value)) {
+                    return true;
+                }
+            } elseif ($emptyValue === $value) {
+                return true;
+            }
+        }
+
+        return false;
     }
 }